public path in pre-loader script setzen
